api/requirements.txt
fastapi==0.95.2
pydantic==1.10.7
requests==2.31.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
